# apps/backend/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Arenda Planform API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /properties:
    post:
      operationId: createProperty
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyCreateRequest'
      responses:
        '201':
          description: Property created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/SubscriptionBlocked'
    get:
      operationId: listProperties
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Properties list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertiesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /properties/archive:
    get:
      operationId: listArchivedProperties
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Archived properties list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertiesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /operations:
    get:
      operationId: listOperations
      security:
        - sessionCookie: []
      parameters:
        - name: type
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/OperationType'
        - name: status
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/OperationStatus'
        - name: category_id
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
              format: uuid
        - name: property_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: recurring_operation_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: lease_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: exclude_archived_properties
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: sort
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/OperationListSort'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
            maximum: 100000
      responses:
        '200':
          description: Operations list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /operations/{id}:
    get:
      operationId: getOperation
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Operation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateOperation
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationUpdateRequest'
      responses:
        '200':
          description: Operation updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/SubscriptionBlocked'
    delete:
      operationId: deleteOperation
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Operation deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/SubscriptionBlocked'
  /operations/{id}/complete:
    post:
      operationId: completeOperation
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Operation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/SubscriptionBlocked'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /operations/{id}/mark-incomplete:
    post:
      operationId: markOperationIncomplete
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Operation marked incomplete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/SubscriptionBlocked'
        '404':
          $ref: '#/components/responses/NotFound'
  /operations/{id}/move:
    post:
      operationId: moveOperation
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationMoveRequest'
      responses:
        '200':
          description: Operation moved to another property of the same owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
      description: >
        The actual cookie name depends on the `cookieSecure` setting:
        `session_id` when insecure and `__Host-session_id` when secure.
...
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        requestId:
          type: string
        code:
          type: string
          description: >-
            Machine-readable error code for failures the client must
            distinguish from the generic HTTP status (e.g.
            `membership_suspended`).
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ProblemError'
    ProblemError:
      type: object
      required: [field, detail]
      properties:
        field:
          type: string
          description: JSON key of the invalid field (e.g. "floor").
        detail:
          type: string
          description: Human-readable explanation of the validation failure.
...
    PropertyAttributes:
      type: object
      additionalProperties: true
      description: >
        Per-type typed property characteristics. Keys and values are in English
        snake_case; Russian labels are a frontend/admin concern. The set of
        valid keys depends on the property type and is validated by the backend
        catalog (see issue #125). An empty object means "no characteristics".
        Only filled keys are stored; null values are never written.
    PropertyCreateRequest:
      type: object
      required: [name, type, address]
      properties:
        name:
          type: string
          maxLength: 255
        type:
          $ref: '#/components/schemas/PropertyType'
        address:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 2000
        attributes:
          $ref: '#/components/schemas/PropertyAttributes'
...
    PropertyAccessContext:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [owner, full_access, viewer]
          description: >
            Actor's role on this property: owner for own properties,
            full_access/viewer for shared ones. Suspended memberships never
            appear here — the detail endpoint answers 403 membership_suspended.
        owner_name:
          type: string
          description: >
            Owner display name ("Name Surname" or masked phone, never email).
            Present only in the detail response when the actor is not the owner.
    PropertyResponse:
      type: object
      required: [id, name, type, address, status, members_count, attributes, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: '#/components/schemas/PropertyType'
        address:
          type: string
        description:
          type: string
        attributes:
          $ref: '#/components/schemas/PropertyAttributes'
        status:
          $ref: '#/components/schemas/PropertyStatus'
        access:
          allOf:
            - $ref: '#/components/schemas/PropertyAccessContext'
        photos:
          type: array
          items:
            $ref: '#/components/schemas/PropertyPhoto'
        members_count:
          type: integer
          minimum: 0
          description: >
            Shared-access participants of the property: membership rows (any
            status, owner excluded — the owner is never a membership row) plus
            pending email invitations.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PropertyPhoto:
      type: object
      required: [id, url]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
...
    PropertiesResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PropertyResponse'
        hidden_shared_count:
          type: integer
          description: >
            Number of shared properties hidden from the recipient due to a
            tariff slot shortage (suspended memberships). Zero for owners and
            when the recipient is within their limit.
          default: 0
    PropertyType:
      type: string
      enum: [apartment, room, apartments, house, commercial, office, warehouse, garage, parking, land]
    PropertyStatus:
      type: string
      enum: [active, maintenance, archived]
...
    OperationType:
      type: string
      enum: [income, expense]
...
    OperationStatus:
      type: string
      enum: [pending, overdue, paid, received, unconfirmed]
    OperationListSort:
      type: string
      enum: [operation_date_desc, operation_date_asc]
      default: operation_date_desc
...
    OperationMoveRequest:
      type: object
      required: [property_id]
      properties:
        property_id:
          type: string
          format: uuid
          description: Target property. Must belong to the same owner as the operation's current property.
    OperationUpdateRequest:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/OperationType'
        category_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 50
        amount_kopecks:
          type: integer
          format: int64
        operation_date:
          type: string
          format: date
        comment:
          type: string
          maxLength: 2000
        lease_id:
          type: string
          format: uuid
        reminder_offset_days:
          type: integer
          nullable: true
          enum: [0, 1, 3, 7]
          description: 0 clears any existing reminder; 1/3/7 reschedules the reminder.
    OperationResponse:
      type: object
      required: [id, owner_id, type, category_id, category_name, name, amount_kopecks, operation_date, status, is_exception, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        property_id:
          type: string
          format: uuid
          nullable: true
          description: Null when the property was deleted in detach mode.
        property_status:
          $ref: '#/components/schemas/PropertyStatus'
        lease_id:
          type: string
          format: uuid
          nullable: true
        recurring_operation_id:
          type: string
          format: uuid
          nullable: true
        type:
          $ref: '#/components/schemas/OperationType'
        category_id:
          type: string
          format: uuid
        category_name:
          type: string
        name:
          type: string
        amount_kopecks:
          type: integer
          format: int64
        operation_date:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/OperationStatus'
        comment:
          type: string
          nullable: true
        is_exception:
          type: boolean
        reminder_offset_days:
          type: integer
          nullable: true
          enum: [1, 3, 7]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OperationsResponse:
      type: object
      required: [items, limit, offset, has_more]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OperationResponse'
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0
        has_more:
          type: boolean
        next_offset:
          type: integer
          minimum: 0
          nullable: true
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    SubscriptionBlocked:
      description: Subscription is blocked; renew to continue mutating data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...